Remove internal imports of ResultsRangeDict
